Draft split Split from Draft.py
